Rename utils.py in tests to utils_for_tests.py
