+++ tests/completor/test_endpoint.py
@@ -5,7 +5,7 @@ import shutil
 from pathlib import Path
 
 import pytest
-from utils import completor_runner
+from utils_for_tests import completor_runner
 
 _TESTDIR_DROGON = Path(__file__).absolute().parent / "data" / "drogon"
