+++ tests/completor/test_error_handling_and_logging.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from zipfile import ZipFile
 
 import pytest
-from utils import completor_runner
+from utils_for_tests import completor_runner
 
 _testdir = Path(__file__).absolute().parent / "data"
 _test_file = "ml_well.sch"
