+++ tests/completor/test_gp_perf.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-import utils
+import utils_for_tests
 
 _TESTDIR = Path(__file__).absolute().parent / "data"
 _TEST_FILE = "test.sch"
@@ -23,8 +23,8 @@ def test_gp_perf_default(tmpdir):
     No GP_PERF_DEVICELAYER keyword set, default to False.
     """
     tmpdir.chdir()
-    utils.open_files_run_create(GP_PERF_BASE_CASE, WELL_DEFINITION, _TEST_FILE)
-    utils.assert_results(WELL_DEFINITION, _TEST_FILE)
+    utils_for_tests.open_files_run_create(GP_PERF_BASE_CASE, WELL_DEFINITION, _TEST_FILE)
+    utils_for_tests.assert_results(WELL_DEFINITION, _TEST_FILE)
 
 
 def test_gp_perf_false(tmpdir):
@@ -40,8 +40,8 @@ def test_gp_perf_false(tmpdir):
      FALSE
     /
     """
-    utils.open_files_run_create(case_file, WELL_DEFINITION, _TEST_FILE)
-    utils.assert_results(WELL_DEFINITION, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, WELL_DEFINITION, _TEST_FILE)
+    utils_for_tests.assert_results(WELL_DEFINITION, _TEST_FILE)
 
 
 def test_gp_perf_true(tmpdir):
@@ -58,8 +58,8 @@ def test_gp_perf_true(tmpdir):
     /
     """
     true_file = Path(_TESTDIR / "wb_perf.true")
-    utils.open_files_run_create(case_file, WELL_DEFINITION, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, WELL_DEFINITION, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_mix_in_branch(tmpdir):
@@ -79,8 +79,8 @@ def test_mix_in_branch(tmpdir):
     /
     """
     true_file = Path(_TESTDIR / "wb_perf_mix_inbranch.true")
-    utils.open_files_run_create(case_file, WELL_DEFINITION, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, WELL_DEFINITION, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_mix_multibranch(tmpdir):
@@ -102,5 +102,5 @@ def test_mix_multibranch(tmpdir):
     with open(Path(_TESTDIR / "welldefinition_2branch.testfile"), encoding="utf-8") as f:
         schedule_file = f.read()
     true_file = Path(_TESTDIR / "wb_perf_mix_multibranch.true")
-    utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
