+++ tests/completor/test_lateral_routing.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 
 import pytest
-import utils
+import utils_for_tests
 
 from completor.exceptions import CompletorError
 
@@ -69,8 +69,8 @@ def test_lat2device(tmpdir):
     """
     schedule_file = Path(_TESTDIR / "ml_well.sch")
     true_file = Path(_TESTDIR / "lat2device.true")
-    utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_lat2device_non_existing(tmpdir):
@@ -106,8 +106,8 @@ LATERAL_TO_DEVICE
     schedule_file = Path(_TESTDIR / "ml_well.sch")
     true_file = Path(_TESTDIR / "lat2device_nonexisting.true")
 
-    utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_lat2device_no_device(tmpdir, caplog):
@@ -149,7 +149,7 @@ LATERAL_TO_DEVICE
     """
     schedule_file = Path(_TESTDIR / "ml_well_l2d_nodevicetest.sch")
     with pytest.raises(CompletorError) as e:
-        utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
     assert "Cannot find a device layer at junction of lateral 2 in A1" in str(e)
 
 
@@ -175,5 +175,5 @@ def test_lat2tubing(tmpdir):
     """
     schedule_file = Path(_TESTDIR / "ml_well.sch")
     true_file = Path(_TESTDIR / "lat2tubing.true")
-    utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
