+++ tests/completor/test_lumping_failure.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-import utils
+import utils_for_tests
 
 _TESTDIR = Path(__file__).absolute().parent / "data"
 _TEST_FILE = "test.sch"
@@ -208,8 +208,8 @@ def test_a1_cells(tmpdir):
     tmpdir.chdir()
     case_file = A1_CELLS_WITH_TOPBOTTOM
     true_file = Path(_TESTDIR / "a1_cells_with_topbottom.true")
-    utils.open_files_run_create(case_file, A1, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, A1, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_a1_lump_with_topbottom(tmpdir):
@@ -219,8 +219,8 @@ def test_a1_lump_with_topbottom(tmpdir):
     tmpdir.chdir()
     case_file = A1_LUMPING_WITH_TOPBOTTOM
     true_file = Path(_TESTDIR / "a1_lumping_with_topbottom.true")
-    utils.open_files_run_create(case_file, A1, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, A1, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_a1_lump_without_topbottom(tmpdir):
@@ -230,8 +230,8 @@ def test_a1_lump_without_topbottom(tmpdir):
     tmpdir.chdir()
     case_file = A1_LUMPING_WITHOUT_TOPBOTTOM
     true_file = Path(_TESTDIR / "a1_lumping_without_topbottom.true")
-    utils.open_files_run_create(case_file, A1, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, A1, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_b1_lump_with_topbottom(tmpdir):
@@ -241,8 +241,8 @@ def test_b1_lump_with_topbottom(tmpdir):
     tmpdir.chdir()
     case_file = B1_LUMPING_WITH_TOPBOTTOM
     true_file = Path(_TESTDIR / "b1_lumping_with_topbottom.true")
-    utils.open_files_run_create(case_file, B1, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, B1, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_b1_lump_without_topbottom(tmpdir):
@@ -253,7 +253,7 @@ def test_b1_lump_without_topbottom(tmpdir):
     print(tmpdir)
     case_file = B1_LUMPING_WITHOUT_TOPBOTTOM
     true_file = Path(_TESTDIR / "b1_lumping_without_topbottom.true")
-    utils.open_files_run_create(case_file, B1, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, B1, _TEST_FILE)
     print(true_file)
     print(_TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
