+++ tests/completor/test_missing_keywords.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 
 import pytest
-import utils
+import utils_for_tests
 
 from completor.constants import Content
 
@@ -131,7 +131,7 @@ def test_minimum_input(tmpdir, capsys):
     _, _outfile, case_file, schedule_file = set_files(tmpdir)
     set_case(Content.PERFORATED, ["completion"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
-    utils.open_files_run_create(case_file, schedule_file, _outfile)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
     captured = capsys.readouterr()
     assert captured.err == ""
     assert captured.out == ""
@@ -143,7 +143,7 @@ def test_missing_welspecs(tmpdir, capsys):
     _, _outfile, case_file, schedule_file = set_files(tmpdir)
     set_case(Content.PERFORATED, ["completion"], case_file)
     set_schedule(["compdat", "welsegs", "compsegs"], schedule_file)
-    utils.open_files_run_create(case_file, schedule_file, _outfile)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
     captured = capsys.readouterr()
     assert captured.err == ""
     assert captured.out == ""
@@ -157,7 +157,7 @@ def test_missing_compdat(tmpdir):
     set_case(Content.PERFORATED, ["completion"], case_file)
     set_schedule(["welspecs", "welsegs", "compsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
 
 
 def test_missing_welsegs(tmpdir):
@@ -168,7 +168,7 @@ def test_missing_welsegs(tmpdir):
     set_case(Content.PERFORATED, ["completion"], case_file)
     set_schedule(["welspecs", "compdat", "compsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
 
 
 def test_inconsistent_files(tmpdir):
@@ -181,7 +181,7 @@ def test_inconsistent_files(tmpdir):
     set_case(Content.PERFORATED, ["completion"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
 
 
 def test_missing_completion(tmpdir):
@@ -192,7 +192,7 @@ def test_missing_completion(tmpdir):
     set_case(Content.PERFORATED, ["wsegaicd"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
 
 
 def test_missing_wsegaicd(tmpdir):
@@ -203,7 +203,7 @@ def test_missing_wsegaicd(tmpdir):
     set_case(Content.AUTONOMOUS_INFLOW_CONTROL_DEVICE, ["completion"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
 
 
 def test_missing_wsegsicd(tmpdir):
@@ -214,7 +214,7 @@ def test_missing_wsegsicd(tmpdir):
     set_case(Content.INFLOW_CONTROL_DEVICE, ["completion"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
 
 
 def test_missing_wsegvalv(tmpdir):
@@ -225,7 +225,7 @@ def test_missing_wsegvalv(tmpdir):
     set_case(Content.VALVE, ["completion"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
 
 
 def test_full_wsegdar(tmpdir, capsys):
@@ -239,7 +239,7 @@ def test_full_wsegdar(tmpdir, capsys):
     _, _outfile, case_file, schedule_file = set_files(tmpdir)
     set_case(Content.DENSITY_ACTIVATED_RECOVERY, ["completion", "wsegdar"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
-    utils.open_files_run_create(case_file, schedule_file, _outfile)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
     captured = capsys.readouterr()
     assert captured.err == ""
     assert captured.out == ""
@@ -253,7 +253,7 @@ def test_missing_wsegdar(tmpdir):
     set_case(Content.DENSITY_ACTIVATED_RECOVERY, ["completion"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
 
 
 def test_full_wsegaicv(tmpdir, capsys):
@@ -267,7 +267,7 @@ def test_full_wsegaicv(tmpdir, capsys):
     _, _outfile, case_file, schedule_file = set_files(tmpdir)
     set_case(Content.AUTONOMOUS_INFLOW_CONTROL_VALVE, ["completion", "wsegaicv"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
-    utils.open_files_run_create(case_file, schedule_file, _outfile)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
     captured = capsys.readouterr()
     assert captured.err == ""
     assert captured.out == ""
@@ -281,4 +281,4 @@ def test_missing_wsegaicv(tmpdir):
     set_case(Content.AUTONOMOUS_INFLOW_CONTROL_VALVE, ["completion"], case_file)
     set_schedule(["welspecs", "compdat", "welsegs", "compsegs"], schedule_file)
     with pytest.raises(ValueError, match=expected_error_message):
-        utils.open_files_run_create(case_file, schedule_file, _outfile)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _outfile)
