+++ tests/completor/test_use_strict.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 
 import pytest
-import utils
+import utils_for_tests
 
 from completor.exceptions import CompletorError
 
@@ -63,8 +63,8 @@ TRUE
     """
     schedule_file = Path(_TESTDIR / "ml_well.sch")
     true_file = Path(_TESTDIR / "usestrict_true.true")
-    utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
 
 
 def test_use_strict_true_false_default(tmpdir):
@@ -98,8 +98,8 @@ USE_STRICT
     schedule_file = Path(_TESTDIR / "ml_well.sch")
 
     true_file_strict_true = Path(_TESTDIR / "usestrict_true.true")
-    utils.open_files_run_create(case_use_strict_true, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file_strict_true, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_use_strict_true, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file_strict_true, _TEST_FILE)
 
     # Test with USESTRICT set to False
     case_use_strict_false = f"""
@@ -110,8 +110,8 @@ USE_STRICT
 {JOINT_LENGTH_AND_WSEGAICD}
 """
     true_file_strict_false = Path(_TESTDIR / "usestrict_false.true")
-    utils.open_files_run_create(case_use_strict_false, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file_strict_false, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_use_strict_false, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file_strict_false, _TEST_FILE)
 
     # USESTRICT keyword not defined
     case_use_strict_default = f"""
@@ -119,8 +119,8 @@ USE_STRICT
 {JOINT_LENGTH_AND_WSEGAICD}
 """
     true_file_strict_default = Path(_TESTDIR / "usestrict_default.true")
-    utils.open_files_run_create(case_use_strict_default, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file_strict_default, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_use_strict_default, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file_strict_default, _TEST_FILE)
 
 
 def test_use_strict_true_missing_branch(tmpdir):
@@ -155,7 +155,7 @@ USE_STRICT
     """
     schedule_file = Path(_TESTDIR / "ml_well.sch")
     with pytest.raises(CompletorError) as e:
-        utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
     assert "USE_STRICT True: Define all branches in case file." in str(e)
 
 
@@ -184,7 +184,7 @@ COMPLETION
     """
     schedule_file = Path(_TESTDIR / "ml_well.sch")
     with pytest.raises(CompletorError) as e:
-        utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+        utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
     assert "USE_STRICT True: Define all branches in case file." in str(e)
 
 
@@ -215,5 +215,5 @@ USE_STRICT
     """
     schedule_file = Path(_TESTDIR / "ml_well.sch")
     true_file = Path(_TESTDIR / "usestrict_false_missingbranch.true")
-    utils.open_files_run_create(case_file, schedule_file, _TEST_FILE)
-    utils.assert_results(true_file, _TEST_FILE)
+    utils_for_tests.open_files_run_create(case_file, schedule_file, _TEST_FILE)
+    utils_for_tests.assert_results(true_file, _TEST_FILE)
