+++ tests/completor/utils_for_tests.py
